UI service has been added to contact with the other services easily and visually

+++ customer-service/src/main/resources/schema.sql
@@ -1,5 +1,5 @@
 CREATE TABLE CUSTOMER (
   id INT AUTO_INCREMENT PRIMARY KEY,
   baseInfoId INT,
-  cifNumber INT
+  cifNumber BIGINT
 );
